+++ src/core/core.ts
@@ -13,8 +13,6 @@ export interface Component<S> {
   logAll?: boolean
   // composition
   components?: Components
-  // child component definitions, used for referencing dynamic components in hot-swaping
-  defs?: Components
   // general purpose groups, used for styles
   groups?: {
     [name: string]: Group,