+++ src/examples/complexList/Root.ts
@@ -20,8 +20,6 @@ export const state = {
   items: {},
 }
 
-export const defs = { Item }
-
 export type S = typeof state
 
 export const inputs: Inputs<S> = ({ stateOf, toIt, toChild, nest, unnest }) => ({